Making the marker component work with the Marker class; moving color calculations into a separate service
